Fixed various cases of misforwardings
Sometimes containers fail to obtain an IP in 10 seconds, or have
an IP in the wrong subnet on a different interface, or have IPv6.
Filter out all those cases, accept only addresses from $LXC_NETWORK.

And wait as long (or as short) as needed for the container to obtain
a usable IP address.